Added Debug/Testing
With this set-up you can see where the allergen test is working / not working. Select some allergens (ice and oj being good common ones) on the webpage, open the console and search a city. The allergen test does find the allergens, however, it only run after the previous API chain is finished. I.e. afterall the drinks have already been pasted to the webpage and it only tests the first drink put in.